Can be used with a specific dbgh5 binary; otherwise a git clone is done and a brand new dbgh5 binary is compiled and used

+++ gatb-core/test/functional/test1/run.sh
@@ -3,6 +3,7 @@
 ################################################################################
 # ARG 1 : login on gforge INRIA
 # ARG 2 : email address to send a report 
+# ARG 3 : dbgh5 binary; if not set, we do a git clone and build the binary 
 ################################################################################
 
 set -e
@@ -21,20 +22,48 @@ else
     export set TEST_MAIL=$2
 fi  
 
+if [ -z "$3" ]; 
+then
+    export set TEST_BINARY=bin/dbgh5
+    export set TEST_GIT_CLONE=1
+else
+    export set TEST_BINARY=$3
+    export set TEST_GIT_CLONE=0
+fi  
+
 
 ################################################################################
-# we clone the git repository
+# we dump the configuration
 ################################################################################
-git clone git+ssh://$TEST_USER@scm.gforge.inria.fr//gitroot/gatb-core/gatb-core.git
 
+echo "----------------------------------------------------------------------"
+echo "                  NON REGRESSION TEST FOR dbgh5                       "
+echo "----------------------------------------------------------------------"
+
+echo "TEST_USER      : " $TEST_USER
+echo "TEST_MAIL      : " $TEST_MAIL
+echo "TEST_BINARY    : " $TEST_BINARY
+echo "TEST_GIT_CLONE : " $TEST_GIT_CLONE
+echo ""
 
 ################################################################################
-# we build the project
+# we may have to clone the git directory if no dbgh5 is provided
 ################################################################################
-cd gatb-core/gatb-core
 
-# we build the project
-mkdir build; cd build; cmake ..; make 
+if [ $TEST_GIT_CLONE -eq 1 ]; then
+    ################################################################################
+    # we clone the git repository
+    ################################################################################
+    git clone git+ssh://$TEST_USER@scm.gforge.inria.fr//gitroot/gatb-core/gatb-core.git
+
+    ################################################################################
+    # we build the project
+    ################################################################################
+    cd gatb-core/gatb-core
+
+    # we build the project
+    mkdir build; cd build; cmake ..; make 
+fi
 
 
 ################################################################################
@@ -51,42 +80,42 @@ wget "ftp://ftp.sra.ebi.ac.uk/vol1/fastq/SRR178/000/SRR1785130/SRR1785130.fastq.
 launch () {
     echo "--------------------------------------------------------- $1  $3  ---------------------------------------------------------"
     
-    bin/dbgh5  -kmer-size $3  -in $1 -check $2
+    $TEST_BINARY  -kmer-size $3  -in $1 -check $2
     
-    bin/dbgh5  -kmer-size $3  -in $1 -check $2 -bloom basic     -debloom original   -debloom-impl basic
-    bin/dbgh5  -kmer-size $3  -in $1 -check $2 -bloom cache     -debloom original   -debloom-impl basic
-    bin/dbgh5  -kmer-size $3  -in $1 -check $2 -bloom neighbor  -debloom original   -debloom-impl basic
-    bin/dbgh5  -kmer-size $3  -in $1 -check $2 -bloom basic     -debloom cascading  -debloom-impl basic
-    bin/dbgh5  -kmer-size $3  -in $1 -check $2 -bloom cache     -debloom cascading  -debloom-impl basic
-    bin/dbgh5  -kmer-size $3  -in $1 -check $2 -bloom neighbor  -debloom cascading  -debloom-impl basic
-    bin/dbgh5  -kmer-size $3  -in $1 -check $2 -bloom neighbor  -debloom original   -debloom-impl minimizer
-    bin/dbgh5  -kmer-size $3  -in $1 -check $2 -bloom neighbor  -debloom cascading  -debloom-impl minimizer
+    $TEST_BINARY  -kmer-size $3  -in $1 -check $2 -bloom basic     -debloom original   -debloom-impl basic
+    $TEST_BINARY  -kmer-size $3  -in $1 -check $2 -bloom cache     -debloom original   -debloom-impl basic
+    $TEST_BINARY  -kmer-size $3  -in $1 -check $2 -bloom neighbor  -debloom original   -debloom-impl basic
+    $TEST_BINARY  -kmer-size $3  -in $1 -check $2 -bloom basic     -debloom cascading  -debloom-impl basic
+    $TEST_BINARY  -kmer-size $3  -in $1 -check $2 -bloom cache     -debloom cascading  -debloom-impl basic
+    $TEST_BINARY  -kmer-size $3  -in $1 -check $2 -bloom neighbor  -debloom cascading  -debloom-impl basic
+    $TEST_BINARY  -kmer-size $3  -in $1 -check $2 -bloom neighbor  -debloom original   -debloom-impl minimizer
+    $TEST_BINARY  -kmer-size $3  -in $1 -check $2 -bloom neighbor  -debloom cascading  -debloom-impl minimizer
     
-    bin/dbgh5  -kmer-size $3  -in $1 -check $2 -nb-cores 1
-    bin/dbgh5  -kmer-size $3  -in $1 -check $2 -nb-cores 2
-    bin/dbgh5  -kmer-size $3  -in $1 -check $2 -nb-cores 4
-    bin/dbgh5  -kmer-size $3  -in $1 -check $2 -nb-cores 8
+    $TEST_BINARY  -kmer-size $3  -in $1 -check $2 -nb-cores 1
+    $TEST_BINARY  -kmer-size $3  -in $1 -check $2 -nb-cores 2
+    $TEST_BINARY  -kmer-size $3  -in $1 -check $2 -nb-cores 4
+    $TEST_BINARY  -kmer-size $3  -in $1 -check $2 -nb-cores 8
 
-    bin/dbgh5  -kmer-size $3  -in $1 -check $2 -max-memory 1000
-    bin/dbgh5  -kmer-size $3  -in $1 -check $2 -max-memory 1000 -nb-cores 1
-    bin/dbgh5  -kmer-size $3  -in $1 -check $2 -max-memory 1000 -nb-cores 2
-    bin/dbgh5  -kmer-size $3  -in $1 -check $2 -max-memory 1000 -nb-cores 4
-    bin/dbgh5  -kmer-size $3  -in $1 -check $2 -max-memory 1000 -nb-cores 8
+    $TEST_BINARY  -kmer-size $3  -in $1 -check $2 -max-memory 1000
+    $TEST_BINARY  -kmer-size $3  -in $1 -check $2 -max-memory 1000 -nb-cores 1
+    $TEST_BINARY  -kmer-size $3  -in $1 -check $2 -max-memory 1000 -nb-cores 2
+    $TEST_BINARY  -kmer-size $3  -in $1 -check $2 -max-memory 1000 -nb-cores 4
+    $TEST_BINARY  -kmer-size $3  -in $1 -check $2 -max-memory 1000 -nb-cores 8
 
-    bin/dbgh5  -kmer-size $3  -in $1 -check $2 -max-memory 1000 -max-disk 2000 
-    bin/dbgh5  -kmer-size $3  -in $1 -check $2 -max-memory 1000 -max-disk 2000 -nb-cores 1
-    bin/dbgh5  -kmer-size $3  -in $1 -check $2 -max-memory 1000 -max-disk 2000 -nb-cores 2
-    bin/dbgh5  -kmer-size $3  -in $1 -check $2 -max-memory 1000 -max-disk 2000 -nb-cores 4
-    bin/dbgh5  -kmer-size $3  -in $1 -check $2 -max-memory 1000 -max-disk 2000 -nb-cores 8
+    $TEST_BINARY  -kmer-size $3  -in $1 -check $2 -max-memory 1000 -max-disk 2000 
+    $TEST_BINARY  -kmer-size $3  -in $1 -check $2 -max-memory 1000 -max-disk 2000 -nb-cores 1
+    $TEST_BINARY  -kmer-size $3  -in $1 -check $2 -max-memory 1000 -max-disk 2000 -nb-cores 2
+    $TEST_BINARY  -kmer-size $3  -in $1 -check $2 -max-memory 1000 -max-disk 2000 -nb-cores 4
+    $TEST_BINARY  -kmer-size $3  -in $1 -check $2 -max-memory 1000 -max-disk 2000 -nb-cores 8
 
-    bin/dbgh5  -kmer-size $3  -in $1 -check $2 -minimizer-size 6
-    bin/dbgh5  -kmer-size $3  -in $1 -check $2 -minimizer-size 7
-    bin/dbgh5  -kmer-size $3  -in $1 -check $2 -minimizer-size 8
-    bin/dbgh5  -kmer-size $3  -in $1 -check $2 -minimizer-size 9
-    bin/dbgh5  -kmer-size $3  -in $1 -check $2 -minimizer-size 10
+    $TEST_BINARY  -kmer-size $3  -in $1 -check $2 -minimizer-size 6
+    $TEST_BINARY  -kmer-size $3  -in $1 -check $2 -minimizer-size 7
+    $TEST_BINARY  -kmer-size $3  -in $1 -check $2 -minimizer-size 8
+    $TEST_BINARY  -kmer-size $3  -in $1 -check $2 -minimizer-size 9
+    $TEST_BINARY  -kmer-size $3  -in $1 -check $2 -minimizer-size 10
 
-    bin/dbgh5  -kmer-size $3  -in $1 -check $2 -minimizer-type 0 
-    bin/dbgh5  -kmer-size $3  -in $1 -check $2 -minimizer-type 1 
+    $TEST_BINARY  -kmer-size $3  -in $1 -check $2 -minimizer-type 0 
+    $TEST_BINARY  -kmer-size $3  -in $1 -check $2 -minimizer-type 1 
 }
 
 # we get the check directory  
@@ -117,8 +146,11 @@ launch ./ERR039477.fastq.gz,./SRR387476.fastq.gz   $TEST_CHECK/k95/ERR039477_SRR
 ################################################################################
 # clean up
 ################################################################################
-cd ../../..
-\rm -rf gatb-core
+
+if [ $TEST_GIT_CLONE -eq 1 ]; then
+    cd ../../..
+    \rm -rf gatb-core
+fi
 
 if [ $? -eq 0 ]; then
